Analytics: Do Not Track disables ad-tracking and Google Analytics
